+++ src/armor-data.ts
@@ -1,0 +1,3 @@
+import { ArmorData } from './types';
+
+export const armorData: ArmorData = require('../armorData.json');
